# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/person": {
            "post": {
                "summary": "Creates a new person",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_megalypse_golang-clean-arch_internal_domain_models.Person"
                        }
                    }
                }
            }
        },
        "/person/filter": {
            "post": {
                "summary": "Filter person",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_megalypse_golang-clean-arch_internal_domain_models.Paginated-github_com_megalypse_golang-clean-arch_internal_domain_models_Person"
                        }
                    }
                }
            }
        },
        "/person/{personId}": {
            "get": {
                "summary": "Gets a person by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_megalypse_golang-clean-arch_internal_domain_models.Person"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_megalypse_golang-clean-arch_internal_domain_models.Paginated-github_com_megalypse_golang-clean-arch_internal_domain_models_Person": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_megalypse_golang-clean-arch_internal_domain_models.Person"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_megalypse_golang-clean-arch_internal_domain_models.Person": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}